project/requirements.txt
# ============================================================
# VisionDriver AI - Full Requirements
# Python 3.10+ recommended
# ============================================================

# --- Core ML & Vision ---
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
ultralytics>=8.2.0
opencv-python>=4.8.0.76
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0
scipy>=1.11.0

# --- Face & Driver Monitoring (Phase 2) ---
mediapipe>=0.10.9
dlib>=19.24.2
imutils>=0.5.4

# --- Emotion / Expression Recognition ---
deepface>=0.0.86
tf-keras>=2.15.0          # Required by deepface

# --- Depth Estimation ---
timm>=0.9.12               # MiDaS / DPT models

# --- Object Tracking ---
supervision>=0.18.0
filterpy>=1.4.5            # Kalman filter (SORT tracker)
lap>=0.4.0                 # Linear assignment (ByteTrack)

# --- OCR (biển báo chữ số, tốc độ) ---
easyocr>=1.7.1
pytesseract>=0.3.10

# --- NLP / QA Engine ---
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
transformers>=4.38.0
langchain>=0.1.16
langchain-community>=0.0.30

# --- Geometry & Signal Processing ---
shapely>=2.0.0
scikit-learn>=1.3.0

# --- Text-to-Speech / Speech-to-Text (tiếng Việt) ---
openai-whisper>=20231117   # STT - speech to text
gTTS>=2.5.0                # TTS - text to speech (online)
pyttsx3>=2.90              # TTS - offline fallback
playsound>=1.3.0

# --- API & Integration Layer ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0

# --- Emergency & Notification ---
twilio>=8.13.0             # Emergency call / SMS
httpx>=0.25.0              # Async HTTP client

# --- Data Pipeline & Training ---
albumentations>=1.3.1      # Data augmentation
wandb>=0.16.5              # Experiment tracking
tensorboard>=2.14.0
datasets>=2.18.0           # HuggingFace datasets
roboflow>=1.1.27           # Dataset download from Roboflow
kagglehub>=0.3.7

# --- Utilities ---
loguru>=0.7.2              # Structured logging
python-dotenv>=1.0.0       # Environment variables
tqdm>=4.66.0               # Progress bars
pyyaml>=6.0.1              # Config files
rich>=13.7.0               # Beautiful terminal output
schedule>=1.2.1            # Task scheduling

# --- Simulation ---
# carla==0.9.14            # Install manually from CARLA release
#   -> See: https://carla.readthedocs.io/en/0.9.14/start_quickstart/

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
